change the way the timer is used with spring
