+++ GirlsFlix/src/main/java/com/gfx/controller/IndexController.java
@@ -15,7 +15,7 @@ import com.gfx.domain.series.Data;
 import com.gfx.domain.series.Genre;
 import com.gfx.domain.series.Serie;
 import com.gfx.domain.users.Enjoyer;
-import com.gfx.service.Scheduler;
+import com.gfx.service.Schedulerbis;
 import com.gfx.service.SerieFactory;
 import com.gfx.service.UserDB;
 import com.gfx.service.UserService;
@@ -30,7 +30,6 @@ class ResourceNotFoundException extends RuntimeException {
 public class IndexController {
 	@Inject
 	SerieFactory serieFactory = new SerieFactory();
-	//Scheduler scheduler = new Scheduler(serieFactory);
 	String message = "Welcome!";
