+++ GirlsFlix/src/main/java/com/gfx/helper/TimerTest.java
@@ -1,0 +1,9 @@
+package com.gfx.helper;
+
+public class TimerTest {
+	
+	public void lancementProgramme() {
+		System.out.println("Lancement");
+	}
+
+}
