+++ GirlsFlix/src/main/webapp/WEB-INF/dispatcher-servlet.xml
@@ -34,4 +34,25 @@
 		</property>
 	</bean>
 	
+	<bean id="tachePlannifie" class="com.gfx.service.Scheduler"></bean>
+	
+	<bean id="tachePlannifieTask" class="org.springframework.scheduling.support.MethodInvokingRunnable">
+		<property name="targetObject"><ref bean="tachePlannifie"/> </property>
+		<property name="targetMethod"><value>taskToExecute</value></property>
+	</bean>
+	
+	<bean id="tachePlannifieScheduledTask" class="org.springframework.scheduling.concurrent.ScheduledExecutorTask">
+		<property name="delay" value="10000"/>
+		<property name="period" value = "30000"/>
+		<property name="runnable" ref="tachePlannifieTask"/>
+	</bean>
+	
+	<bean id="timerFactory" class="org.springframework.scheduling.concurrent.ScheduledExecutorFactoryBean">
+		<property name="scheduledExecutorTasks">
+			<list>
+				<ref bean="tachePlannifieScheduledTask"/>
+			</list>
+		</property>
+	</bean>
+	
 </beans>
